Fixed bug in MCTS with rollout policy
This now puts the acting players hand into the deck before makign a decision, and then removes it once decision is made, and before the action is applied to the gamestate.
Otherwise we have a problem with the player never actually thinking they have the actual cards that they do have!